Work on CIS
Worked on CIS research, starting units, and AI. Removed the Heavy Iron Division as the AI template designer had no freakin clue what it was looking at. I could have added it as a legitimate design in an ai_template file for CIS, but the division design itself was rather unnecessary, so I removed it completely. Also removed their designs being the same as ROC's, because was creating, only for CIS, weird templates of only light and medium armor without any motorized or any other infantry to speak of.

Also made the CIS AI put more troops to garrison itself, as it previously left itself wide upon far too frequently, given how small of  a nation it is, and how easy it would be to cap all its victory points if there was a limited garrison present.